// src/types/EventType.ts
export type EventTypeUserRequest = EventTypeRequest & {
    user?: string;
};

export type EventTypeOrganisationRequest = EventTypeRequest & {
    organization?: string;
};

type EventTypeRequest = {
    active?: boolean;
    count?: number;
    page_token?: string;
    sort?: string;
};